+++ src/CMakeLists.txt
@@ -1,1 +1,1 @@
-add_subdirectory(utils)
+add_subdirectory(lib)
