+++ src/lib/CMakeLists.txt
@@ -1,0 +1,1 @@
+include_directories(array)
