+++ src/lib/array/CMakeLists.txt
@@ -3,10 +3,10 @@
 # This file sets up the utils static library
 ################################################
 
-add_library(libutils STATIC
+add_library(libarray STATIC
         KArray.cpp
         KArray.hpp
 )
-target_compile_options(libutils PRIVATE
+target_compile_options(libarray PRIVATE
         ${KERN_COMPILER_FLAGS}
 )
